backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.129.1",
    "httpx>=0.27.0",
    "numpy>=2.2.0",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "scipy>=1.15.0",
    "sqlalchemy>=2.0.46",
    "sqlmodel>=0.0.36",
    "google-api-python-client>=2.130.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "greenlet>=3.3.2",
]
